% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_flow.R
\name{CalculateHighestTopicCosineSimilarity}
\alias{CalculateHighestTopicCosineSimilarity}
\title{Calculate Maximimum Likehood Topic}
\usage{
CalculateHighestTopicCosineSimilarity(ttm1, ttm2)
}
\arguments{
\item{ttm1}{A topic term matrix}

\item{ttm2}{A topic term matrix (e.g. of the following month)}
}
\value{
Returns a cartesian product data.table of the ttm pair containing the topics with maximum likehood.
}
\description{
Accepts two topic term matrices generated by different LDA runs.
Topics may have different terms in their respective vocabulary.
Non-existent terms will be added with probability 0 to each other.
See \code{\link{CalculateTopicFlow}} for an example of it's usage.
}
