% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{CalculateLDAModelsInKSetOld}
\alias{CalculateLDAModelsInKSetOld}
\title{Calculate LDA Models up to N Old}
\usage{
CalculateLDAModelsInKSetOld(dfm, Ks, method = "VEM")
}
\arguments{
\item{dfm}{A document frequency matrix}

\item{Ks}{The number of topics (i.e. 1:N).}
}
\value{
A list of models from 1 to N specified on param N.
}
\description{
Legacy function for backward compatibility. Will be removed in the future.
}
