% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{CalculateLDAModelsUpToN}
\alias{CalculateLDAModelsUpToN}
\title{Calculate LDA Models up to N}
\usage{
CalculateLDAModelsUpToN(dfm, N)
}
\arguments{
\item{dfm}{A document frequency matrix}

\item{N}{The number of topics (i.e. 1:N).}
}
\value{
A list of models from 1 to N specified on param N.
}
\description{
Returns a list, where every position is a created model. List starts on k=2.
For more control on the values of k, see \code{\link{CalculateLDAModelsInKSet}}
}
