% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_flow.R
\name{CalculateTopicFlow}
\alias{CalculateTopicFlow}
\title{Calculate Topic Flow}
\usage{
CalculateTopicFlow(models)
}
\arguments{
\item{models}{A list of models. See \code{\link{rawToLDA}}.}
}
\value{
A topic flow data table.
}
\description{
Creates a topic flow table connecting various monthly topics.
}
