% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{ChooseKLDAModelsPerplexity}
\alias{ChooseKLDAModelsPerplexity}
\title{Choose K through Perplexity}
\usage{
ChooseKLDAModelsPerplexity(lda.model.list, topic.size.median.threshold)
}
\arguments{
\item{lda.model.list}{A list of topic models. See \code{\link{CalculateLDAModelsUpToN}}.}
}
\value{
a single integer indicating the number of topics
}
\description{
This function attemps to quantify our visual perception of the elbow method to pick k, the number of topics.
}
