% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{cosinePCA}
\alias{cosinePCA}
\title{Calculate LDAvis Topic Cosine Similarity}
\usage{
cosinePCA(phi)
}
\arguments{
\item{phi, }{as specified in createJSON on LDAvis.}
}
\description{
This function is used as a replacement to `jsPCA` in LDAVis package function createJSON, as
the default can't handle words with probability 0. This also ensures consistency to the different
month topics, as it now uses the same cosine similarity instead of symetric Dkl. For more details,
see issue: https://github.com/sailuh/perceive/issues/84
}
