% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{exportDocumentTermMatrix}
\alias{exportDocumentTermMatrix}
\title{Export Document Term Matrix}
\usage{
exportDocumentTermMatrix(models, exportPath)
}
\arguments{
\item{models}{The models object as output from \code{\link{rawToLDA}}.}

\item{exportPath}{A folder path to export the Document Term Matrix of each month in models}
}
\description{
Exports a document term matrix to a specified *folder* path.
}
