% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{FilterDFMByTFIDF}
\alias{FilterDFMByTFIDF}
\title{Filter DFM by TFIDF}
\usage{
FilterDFMByTFIDF(dfm_quanteda, threshold)
}
\arguments{
\item{dfm_quanteda}{A dfm as defined by the quanteda package}

\item{threshold}{The threshold to filter terms}
}
\value{
A dfm with filtered terms.
}
\description{
Uses threshold to select the top terms based on tf-idf.
}
