% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{GetDocumentsAssignedToTopicK}
\alias{GetDocumentsAssignedToTopicK}
\title{Get Documents Assigned to Topic K}
\usage{
GetDocumentsAssignedToTopicK(lda.model, k)
}
\arguments{
\item{lda.model}{A single lda model.}

\item{k}{The topic k (e.g. topic 5 from a lda model of 10 topics) to list the documents.}
}
\value{
The list of documents for topic k in lda.model.
}
\description{
Note the documents are defined as a distribution over topics.
This returns the highest probability for a given topic,
but it is possible the 2nd to high topic may have about the same probability.
}
