% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{GetDocumentsOfTopTopicK}
\alias{GetDocumentsOfTopTopicK}
\title{Get Documents of Topic K}
\usage{
GetDocumentsOfTopTopicK(lda.model, k)
}
\arguments{
\item{lda.model}{A single lda model.}

\item{k}{The topic k (e.g. topic 5 from a lda model of 10 topics) to list the documents.}
}
\value{
The list of document names.
}
\description{
Returns a vector of characters, where each element is the name of a document associated to topic K.
Note that this can be HIGHLY misleading since a document have a distribution over topics and the top topic
may have a very close probability to occur as the other topic.
}
