% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{GetDocumentTopicMatrices}
\alias{GetDocumentTopicMatrices}
\title{Get Document Topic Matrices}
\usage{
GetDocumentTopicMatrices(lda.model.list)
}
\arguments{
\item{The}{list of models. See \code{\link{rawToLDA}}.}
}
\value{
The ordered document topic matrices.
}
\description{
Returns the topic term matrices for a list of lda models. See also \code{\link{GetTopicTermMatrices}}.
}
