% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{GetTopicTermMatrices}
\alias{GetTopicTermMatrices}
\title{Get Topic Term Matrices}
\usage{
GetTopicTermMatrices(lda.model.list, n.terms = 10)
}
\arguments{
\item{lda.model.list}{The list of lda models. See \code{\link{rawToLDA}}.}

\item{n.terms}{The number of terms to display per topic (note the vocabulary can go up to over 95k for a given month).}
}
\value{
The ordered topic term matrices.
}
\description{
Returns the topic term matrices for a list of lda models. See also \code{\link{GetDocumentTopicMatrices}}.
}
