% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{isSameTopicAndSameCVE}
\alias{isSameTopicAndSameCVE}
\title{Validate if Topics Group or not emails from same CVE ID}
\usage{
isSameTopicAndSameCVE(models, cves, chosen_year)
}
\arguments{
\item{chosen_year}{The year to validate the method, as used in the topic_flow_creation vignette.}

\item{A}{list of lda models. See \code{\link{rawToLDA}}.}

\item{A}{list of cves and file emails, as parsed from CVE Mitre Website. See the validation vignette (cve_mitre_xml_parser_to_yearly_validation.Rmd) for details.}
}
\value{
A table containing the email of the files, their cve id, and wether they mapped or not to the same topic.
}
\description{
This function is used to validate the models from PERCEIVE, by checking if emails
discussing about the same vulnerability were mapped to the same topic (good), or not (bad).
See the topic_flow_creation vignette for it's usage.
}
