% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{loadFiles}
\alias{loadFiles}
\title{Load Corpus Files}
\usage{
loadFiles(raw.corpus.folder.path)
}
\arguments{
\item{raw.corpus.folder.path}{The path to the corpus folder (e.g. 2012.parsed)

Returns a folder used by \code{\link{rawToLDA}}.}
}
\description{
Used to load the files into memory. Assume the format of the new crawler, where each year of
mailing list is inside a folder, and months inside sub-folders.
See \code{\link{rawToLDA}} to see it's usage.
}
\details{
TODO: Parameterize the file extension (currently assumes reply.body.txt)
}
