% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotLDAVis}
\alias{plotLDAVis}
\title{Localhost Interactive Topic Visualization}
\usage{
plotLDAVis(model, as.gist = FALSE, topicSimilarityMethod = jsPCA)
}
\description{
This function is a temporary placeholder for the LDAvis package to test
different cosine similarity measurements.
}
