% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{PlotTermTFIDF}
\alias{PlotTermTFIDF}
\title{Plot Term TF-IDF}
\usage{
PlotTermTFIDF(dfm_quanteda, filter.by.term = FALSE,
  n.terms.threshold = 5000)
}
\arguments{
\item{dfm_quanteda}{A Document Frequency Matrix as defined in the quanteda package.}

\item{filter.by.term}{If TRUE n.terms.threshold will be used.}

\item{n.terms.threshold}{Specify a limit to the number of terms plotted.}
}
\value{
A plot of the terms TF-IDF on the console.
}
\description{
Method was replicated from topicmodels publication on the
Journal of Statistical Software by Grun and Hornik, 2011.
For more details see \code{\link{PrintTermTFIDF}}
}
