% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{PrintTermTFIDF}
\alias{PrintTermTFIDF}
\title{Print Terms TF-IDF}
\usage{
PrintTermTFIDF(dfm_quanteda)
}
\arguments{
\item{dfm_quanteda}{A Document Frequency Matrix as defined in the quanteda package.}
}
\value{
Prints the terms TF-IDF on the console.
}
\description{
Method was replicated from topicmodels publication on the
Journal of Statistical Software by Grun and Hornik, 2011.
This function prints the terms tf-idf values in quartiles.
The rule of thumb used in the journal is to use the value immediately before the mean when
deciding the filtering threshold for \code{\link{FilterDFMByTFIDF}}.
}
