% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{PrintTopicsTopTerm}
\alias{PrintTopicsTopTerm}
\title{Prints Topics Top Term}
\usage{
PrintTopicsTopTerm(lda.model.list)
}
\arguments{
\item{lda.model.list}{A list of lda models. See \code{\link{rawToLDA}}.}
}
\description{
Calls \code{\link{GetDocumentsPerTopicCount}} for every topic. Best used just after creating the LDA models
to get a sense of them for every value of k. Beware this can be HIGHLY misleading since a topic is a distribution
over terms, and the top term may have a very close probability to occur as other term.
}
