% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{rawToLDA}
\alias{rawToLDA}
\title{Raw to LDA}
\usage{
rawToLDA(folder, k, months)
}
\arguments{
\item{folder}{Folder variable containing all raw data. See \code{\link{loadFiles}}.}

\item{k}{The number of topics k for the model.}

\item{months}{A character vector containing the months of interest (currently one uses the first month)}
}
\value{
A list structure containing 'tokens', 'dfm', 'LDA', used by various functions in the package.
}
\description{
Raw to LDA make several assumptions on pre-processing rules. See the topic_flow_creation vignette for more details.
}
