% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{testConceptSimilarity}
\alias{testConceptSimilarity}
\title{Export Leximancer to Topic Flow}
\usage{
testConceptSimilarity(conceptMapYear1 = "~/Desktop/ttm_leximancer_2009.csv",
  conceptMapYear2 = "~/Desktop/ttm_leximancer_2010.csv",
  savePath = "~/Desktop/leximancer_FD_similarity_highest_likelehood_term_representation_2009_2010.csv")
}
\arguments{
\item{conceptMapYear1}{A concept versus concept matrix from Leximancer on a given year.}

\item{conceptMapYear2}{A concept versus concept matrix from Leximancer on a *following* year.}

\item{savePath}{The csv file to save the the topic flow file.}
}
\description{
A transformation is provided for the equivalent of a topic-term matrix in order
that replaces LDAVis. The code below provides a transformation for it's equivalent to topic-term matrix in order
to construct topic flows in the same way as done as LDA.
This section, although functional, is still undergoing work as the tool as only recently acquired by the group.
}
