% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{CalculateLDAModelsInKSet}
\alias{CalculateLDAModelsInKSet}
\title{Create Topic Models from 1 to k}
\usage{
CalculateLDAModelsInKSet(dfm, Ks, method = "VEM")
}
\arguments{
\item{dfm}{A document frequency matrix}

\item{Ks}{A vector of integers (e.g. c(1,3,2,7)) specifying the number of topics to be created}
}
\value{
A list of the models for all chosen values in Ks
}
\description{
Returns a list, where every position is a created model. List of k values must be provided as a vector.
}
